backend/patient-service/requirements.txt
acres==0.2.0
alembic==1.12.0
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
bcrypt==4.0.1
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.1.0
cryptography==41.0.4
distro==1.9.0
dnspython==2.7.0
email-validator==2.0.0
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.17.0
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Cors==4.0.0
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
fpdf==1.7.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
load-dotenv==0.1.0
looseversion==1.3.0
lxml==5.3.0
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.20.1
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.2
openai==1.60.2
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathlib==1.0.1
pillow==11.1.0
pluggy==1.5.0
prometheus-flask-exporter==0.22.4
prometheus_client==0.21.1
prov==2.0.1
psutil==7.0.0
psycopg2-binary==2.9.9
puremagic==1.28
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
pydot==3.0.4
PyJWT==2.8.0
PyMuPDF==1.25.2
pyparsing==3.2.1
PyPDF2==3.0.1
pytesseract==0.3.13
pytest==7.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
pyxnat==1.6.2
rdflib==6.3.2
redis==5.2.1
requests==2.32.3
scipy==1.15.1
sentry-sdk==1.32.0
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.22
tomli==2.2.1
tqdm==4.67.1
traits==7.0.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
